Refactored single core module to multiple material, bp and core modules
